// src/domains/person/interfaces/PersonSchemaFields.ts
import {Types} from "mongoose";
import type {ISO3166Alpha2CountryCode} from "../../../shared/schema/enums/ISO3166Alpha2CountryCodeSchema.js";
import type {CloudinaryImageObject} from "../../../shared/schema/cloudinary/CloudinaryImageObjectSchema.js";

/**
 * Represents a person with basic identifying information,
 * biography, date of birth, nationality, and an optional profile image.
 */
export interface PersonSchemaFields {
    /**
     * Unique identifier for the person.
     * Typically generated by MongoDB.
     */
    readonly _id: Types.ObjectId;

    /**
     * Full name of the person.
     */
    name: string;

    /**
     * A short biography or description of the person.
     */
    biography: string;

    /**
     * Date of birth of the person.
     */
    dob: Date;

    /**
     * Nationality of the person, represented as an ISO 3166-1 alpha-2 code.
     */
    nationality: ISO3166Alpha2CountryCode;

    /**
     * Optional profile image of the person.
     * Can be `null` if the person does not have a profile image.
     */
    profileImage?: CloudinaryImageObject | null;

    slug: string;
}